refactor: split settings into components
